pyproject.toml
[tool.poetry]
name = "yatracker-provider"
version = "0.1.0"
description = ""
authors = ["Rostislav Aleev <rs.aleev@gmail.com>"]
readme = "README.md"
packages = [{include = "yatracker_provider"}]

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.31.0"


[tool.poetry.group.dev.dependencies]
apache-airflow = "^2.6.1"


[tool.poetry.group.test.dependencies]
pytest = "^7.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
